+++ front-end/src/component/DetailPage/Detail.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { UserHeader } from "../Header";
+import { UserHeader } from "../UserHeader";
 import { useLocation, useNavigate } from "react-router-dom";
 import { BlocksRenderer } from "@strapi/blocks-react-renderer";
 import { Button, Form, Select, message } from "antd";
@@ -85,15 +85,15 @@ export default function Detail() {
       .then(() => {
         isAuthenticated
           ? navigate("/transaction", {
-              state: {
-                data: data,
-                Title: Title,
-                Price: totalPrice * count,
-                selectedDate: selectedDate,
-                people: count,
-                packageId: documentId,
-              },
-            })
+            state: {
+              data: data,
+              Title: Title,
+              Price: totalPrice * count,
+              selectedDate: selectedDate,
+              people: count,
+              packageId: documentId,
+            },
+          })
           : navigate("/login");
       })
       .catch((err) => {
